backend/requirements.txt
Django>=4.2.7
djangorestframework>=3.12.4,<4.0
psycopg2>=2.9.9
drf-spectacular>=0.20.0
djangorestframework-simplejwt
paddleocr
paddlepaddle
setuptools>=65.5.1
wheel
opencv-python-headless
numpy>=1.20
spacy
django-filter
pyotp
qrcode[pil]
